Add files via upload
Co-authors Erica EM Moodie, Susan M Shortreed, Christel Renoux. Abstract of JSM talk (2022) in Washington DC:

The sequential treatment decisions made by physicians to treat chronic diseases are formalized in the statistical literature as dynamic treatment regimes. To date, methods for dynamic treatment regimes have been developed under the assumption that observation times are determined by study investigators. That assumption is often not satisfied in electronic health records data in which the outcome, the observation times and the treatment mechanism are driven by patients' characteristics. The treatment and observation processes can lead to spurious associations between the treatment of interest and the outcome to be optimized. We address these associations by incorporating two inverse weights that are functions of patient's covariates into dynamic weighted ordinary least squares to develop optimal one-stage dynamic treatment regimes. The method is used to develop an optimal dynamic treatment regime that chooses between two antidepressants to optimize a utility function related to the change in body mass index. We find a small benefit in using our proposed method to tailor treatment to individuals, as compared to estimators that do not account for the irregular observation times.